backend/pyproject.toml
[project]
name = "exhibition-platform"
version = "0.1.0"
description = "Exhibition event management platform with time-based filtering"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "anthropic>=0.7.0",
    "pillow>=10.1.0",
    "email-validator>=2.1.0",
    "qrcode>=7.4.2",
    "aiofiles>=23.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "models",
    "routes",
    "services",
]
include = [
    "main.py",
    "database.py",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]
